bird-observation-system/experiments/requirements.txt
ultralytics==8.3.74
opencv-python==4.11.0.86
numpy==2.1.1
pandas==2.2.3
matplotlib==3.10.1
PyYAML==6.0.2